Coded 2 Daily programs from reddit
I thought I gave myself time to program 2 intermediate challenges.  I
was going to use the Decoder challenge as part of one of the challenges.
It took me about 5 minutes to program and comment the Decoder challenge.
And about the same for the Easter Challenge, I added Thanksgiving to
make a little more challenging.